Notification: rename endpoint to nofiticationConfigEndpoint
