+++ notification/src/api/rest/notificationConfigEndpoint.ts
